requirements.txt
python-dotenv
langchain
langchain_openai
langchain-anthropic
gradio
chromadb
bs4
